# Improve long feature fetch map draw