@reflectored 2023-04-13
cppcheck failed unrelated to my PR:
> /home/runner/work/QGIS/QGIS/src/providers/hana/qgshanaproviderconnection.cpp:50,style,unsignedLessThanZero,Checking if unsigned expression 'i' is less than zero.