@m-kuhn 2023-04-23
@nyalldawson @PeterPetrik what are your opinions (refs https://github.com/qgis/QGIS/pull/41198)